Restructuring, Advanced Search V2 - Major restructuring has occurred. - Removed empty last line in app.js. - Added card.js, deck.js, and features.js as imports in api.js. - Restructured ALL web queries to utilize new structure. - Added requests/card.js - all card-related functions have been moved here. - Added requests/deck.js - all deck-related functions have been moved here. - Added requests/features.js - all feature-related functions have been moved here. - Added routes/auth.js to project. - Advanced Search V2 is here! - Added support for multi-colors search. - Added support for multi-color_identity search. - Added support for legalities search. - Multi-colors search is a 'contains any' type strictness. - Multi-color_identity search is a 'exact' type strictness. - Legality search can only filter by one legality type and ruling, for now.
