+++ routes/auth.js
@@ -1,0 +1,7 @@
+// Initial Setup
+var express = require('express');
+var router = express.Router();
+
+
+// Must ALWAYS Be At The Bottom
+module.exports = router;
